refactor: container moved to the required folder
