+++ src/widgets/PostList/PostList.tsx
@@ -1,6 +1,6 @@
 import { PostCard } from "../../entities/post/ui/PostCard"
 import { postsData } from "../../entities/post/model/postsData"
-import { Container } from "../../shared/ui/Container"
+import { Container } from "../../shared/ui/Container/Container"
 
 export const PostList = () => {
 	return (
